backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gspread==5.11.3
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
oauth2client==4.1.3
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
